refactory directory
