+++ package.json
@@ -29,10 +29,10 @@
         "cookie-parser": "^1.4.6",
         "cors": "^2.8.5",
         "dotenv": "^16.0.1",
+        "editly": "^0.13.0",
         "express": "^4.18.1",
         "express-session": "^1.17.3",
         "ffmpeg": "^0.0.4",
-        "fluent-ffmpeg": "^2.1.2",
         "jsonwebtoken": "^8.5.1",
         "moment": "^2.29.4",
         "multer": "^1.4.5-lts.1",
