+++ src/middlewares/s3.middleware.js
@@ -1,13 +1,7 @@
 import multer from "multer";
 import multerS3 from "multer-s3";
 import aws from "aws-sdk";
-import dotenv from "dotenv";
-import fluent_ffmpeg from "fluent-ffmpeg";
-import fs from "fs";
 
-dotenv.config();
-
-// s3 접근하기 위한 키 설정
 const s3 = new aws.S3({
     credentials: {
         accessKeyId: process.env.AWS_ID,
@@ -30,119 +24,3 @@ export const video_upload = multer({
     },
     storage: s3_videoUploader,
 });
-
-export const merge_videos = async (video_one, video_two, video_three, kakao_id) => {
-    return new Promise((resolve, reject) => {
-        const concatMP4FileTmpPath = "./tmp";
-        const concatMP4FilePath = `./src/combine/${kakao_id}.mp4`; //합쳐질 파일 위치,이름
-        const targetFiles = [video_one, video_two, video_three];
-
-        let merged_video = fluent_ffmpeg();
-
-        targetFiles.forEach((element) => {
-            //목록 추가하기
-            merged_video = merged_video.addInput(element);
-        });
-
-        //파일 1개로 만들기
-        merged_video
-            .mergeToFile(concatMP4FilePath, concatMP4FileTmpPath)
-            .on("error", (error) => {
-                console.log(error);
-                resolve();
-                throw error;
-            })
-            .on("end", () => {
-                resolve();
-            });
-    });
-};
-
-export const create_video_s3_objects = (video_one, video_two, video_three) => {
-    if (video_one !== undefined && video_two === undefined && video_three === undefined) {
-        const objects = {
-            Bucket: "healthree",
-            Delete: {
-                Objects: [
-                    {
-                        Key: `videos/${video_one}`,
-                    },
-                ],
-            },
-        };
-        return objects;
-    }
-
-    if (video_one !== undefined && video_two !== undefined && video_three === undefined) {
-        const objects = {
-            Bucket: "healthree",
-            Delete: {
-                Objects: [
-                    {
-                        Key: `videos/${video_one}`,
-                    },
-                    {
-                        Key: `videos/${video_two}`,
-                    },
-                ],
-            },
-        };
-        return objects;
-    }
-
-    if (video_one !== undefined && video_two !== undefined && video_three !== undefined) {
-        const objects = {
-            Bucket: "healthree",
-            Delete: {
-                Objects: [
-                    {
-                        Key: `videos/${video_one}`,
-                    },
-                    {
-                        Key: `videos/${video_two}`,
-                    },
-                    {
-                        Key: `videos/${video_three}`,
-                    },
-                ],
-            },
-        };
-        return objects;
-    }
-};
-
-export const delete_video_s3 = (object) => {
-    s3.deleteObjects(object, (error, data) => {
-        if (error) {
-            throw error;
-        }
-    });
-    return;
-};
-
-export const read_video = async (kakao_id) => {
-    const file = fs.createReadStream(`./src/combine/${kakao_id}.mp4`, { flags: "r" });
-
-    return file;
-};
-
-export const s3_upload = async (file) => {
-    const params = {
-        Bucket: "healthree/videos",
-        Key: String(new Date().getTime() + Math.random()),
-        ACL: "public-read", // 권한: 도메인에 객체경로 URL 을 입력하여 접근 가능하게 설정
-        Body: file,
-        ContentType: "video/mp4",
-    };
-    const s3_upload = await s3.upload(params).promise();
-
-    return s3_upload;
-};
-
-export const delete_file = async (file_name) => {
-    fs.unlink(file_name, (error) => {
-        if (error) {
-            throw error;
-        }
-    });
-};
