+++ src/routes/index.js
@@ -1,10 +1,12 @@
 import express from "express";
 const router = express.Router();
 
-import main_router from "./main.routes.js";
-import users_router from "./users.routes.js";
+import video_router from "./video.routes.js";
+import goal_router from "./goal.routes.js";
+import auth_router from "./auth.routes.js";
 
-router.use("/main", main_router);
-router.use("/users", users_router);
+router.use("/videos", video_router);
+router.use("/goals", goal_router);
+router.use("/auth", auth_router);
 
 export default router;
